requirements.txt
Flask
requests
datetime
paho-mqtt
flask-socketio
flask-sqlalchemy
